文件夹下的所有文件都进入回收站

+++ src/main/java/com/simpleclouddisk/domain/dto/FilePageDto.java
@@ -11,7 +11,7 @@ public class FilePageDto {
     /**
      * 父级id
      */
-    private Integer pid;
+    private Long pid;
     /**
      * 当前页
      */
